fix(mq): 修复注册来源ID字段类型错误
- 将SourceID字段类型从int32修改为int64以支持更大的数值范围
- 保持JSON序列化标签和注释不变，确保数据兼容性

+++ mq/model.go
@@ -129,7 +129,7 @@ type UserRegisterNotify struct {
 	DeviceEndpoint string `json:"device_endpoint"` // 设备终端 APP H5 PC
 	RegisterType   int32  `json:"register_type"`   // 注册方式
 	SourceTarget   int32  `json:"source_target"`   // 注册来源 1-活动 2-游戏 3-代理推广
-	SourceID       int32  `json:"source_id"`       // 注册来源id
+	SourceID       int64  `json:"source_id"`       // 注册来源id
 	CountryCode    string `json:"country_code"`    // 国家码
 	RegisterAt     int64  `json:"register_at"`     // 注册时间
 }
